Apply JavaScript for dynamic filters
This sets event handlers on the input and checkbox elements so that when
their values change the list of countries is re-filtered.

It relies on the presence of a red-list/amber-list/green-list within a
country list item for determining that rag status. It uses the name
within the link to determine the text name of a country.