Add JS for filtering and anchor links
This configures an event listener on the country input, when the value
of this changes countries are shown/hidden accordingly. If there are no
countries visible in a colour category then the category is not shown.

I noticed that this filtering broke the anchor links to individual lists
so I added some JS which clears the filter if you click one of these
anchors. I'm not sure if this is an ideal experience but seemed
preferable to doing nothing where the anchors would take you to an empty
space.